Wing bending mode estimator: Add function gyro2velocity
In order to get the first time-derivative of the first
wing bending mode eta_d, distributed gyros can be used.
For the mapping gyros -> eta_d, the rotation angle of
the mode shape matrix was used.
In this commit a second approach is implemented for this
mapping: gyros -> z_d -> eta_d, where z_d is the
vertical velocity. The mapping gyros -> z_d can be
computed using the new function gyro2velocity. The
mapping z_d -> eta_d can be performed using the vertical
displacement of the mode shape matrix (as for
accelerations -> eta_dd). The vertical velocity z_d
could be used for gyro bias estimation in the future.
However, this approach assumes that the local rotation
of the wing is always parallel to the wing bending
curve, which is not the case for the current model
of the soft Funray wing spar. The new method was
implemented in the Simulink library block, but the
old method is still used.